fixed a bug caused by round off errorr

+++ src/Metagene.cpp
@@ -1,5 +1,5 @@
 /*
-* Metagene: Determine the normazlied gene locations for metagene plots 
+* Metagene: Determine the normazlied gene locations for metagene plots
 * Copyright (C) 2025 Rishvanth Prabakar
 *
 * This program is free software; you can redistribute it and/or modify
@@ -30,11 +30,11 @@ using std::unordered_map;
 
 
 void
-Metagene::add_region(const GenomicRegion& g, const vector<string>& fields, 
+Metagene::add_region(const GenomicRegion& g, const vector<string>& fields,
            vector<FeatureRegions>& feature) {
 
-  feature.push_back(FeatureRegions()); 
-  feature.back().chrom = g.name; 
+  feature.push_back(FeatureRegions());
+  feature.back().chrom = g.name;
   feature.back().start = g.start;
   feature.back().end = g.end;
   feature.back().name = fields[0];
@@ -54,23 +54,23 @@ Metagene::process_feature(const vector<FeatureRegions>& feature) {
   size_t feature_size = 0;
   // cout << "Number of regions: " << feature.size() << endl;
   for (size_t i = 0; i < feature.size(); ++i) {
-    // cout << feature[i].chrom << "\t" 
+    // cout << feature[i].chrom << "\t"
     //      << feature[i].start << "\t" << feature[i].end << "\t"
     //      << feature[i].name << "\t" << feature[i].dir << endl;
 
     feature_size += (feature[i].end - feature[i].start);
   }
-  
+
   // cout << "feature size: " << feature_size << endl;
-  float base_pct = static_cast<float>(n_divisions) / 
-                    static_cast<float>(feature_size);
+  double base_pct = static_cast<double>(n_divisions) /
+                    static_cast<double>(feature_size);
   // cout << "base_size: " << base_pct << endl;
 
-  float base_counter = 0;
+  double base_counter = 0;
   if (!feature[0].dir)
     base_counter = 100;
- 
-   
+
+
 
   for (size_t i = 0; i < feature.size(); ++i) {
     size_t prev_counter = static_cast<size_t>(round(base_counter));
@@ -81,7 +81,7 @@ Metagene::process_feature(const vector<FeatureRegions>& feature) {
       metagene.add(feature[i].chrom, j, j + 1,
         FeatureVector<pair<string, size_t>>
         (make_pair(feature[i].name, static_cast<size_t>(round(base_counter)) )) );
-      cout << j << "\t" << base_counter << "\t" 
+      cout << j << "\t" << base_counter << "\t"
            << static_cast<size_t>(round(base_counter)) << endl;
       */
 
@@ -91,56 +91,56 @@ Metagene::process_feature(const vector<FeatureRegions>& feature) {
       else {
         base_counter -= base_pct;
       }
-      
+
       size_t curr_counter = static_cast<size_t>(round(base_counter));
       if (curr_counter != prev_counter) {
         metagene.add(feature[i].chrom, prev_start, j + 1,
           FeatureVector<pair<string, size_t>>
           (make_pair(feature[i].name, prev_counter )) );
-        // cout << "Adding: " << prev_start << " " << j + 1 << " " 
+        // cout << "Adding: " << prev_start << " " << j + 1 << " "
         //      << prev_counter << endl;
         prev_counter = curr_counter;
         prev_start = j + 1;
       }
 
-    }   
-    /* 
+    }
+    /*
     metagene.add(feature[i].chrom, feature[i].end, feature[i].end + 1,
       FeatureVector<pair<string, size_t>>
       (make_pair(feature[i].name, static_cast<size_t>(round(base_counter)) )) );
-    cout << feature[i].end << "\t" << base_counter << "\t" 
+    cout << feature[i].end << "\t" << base_counter << "\t"
          << static_cast<size_t>(round(base_counter)) << endl;
     */
     metagene.add(feature[i].chrom, prev_start, feature[i].end + 1,
       FeatureVector<pair<string, size_t>>
-      (make_pair(feature[i].name, prev_counter )) );    
-    // cout << "Adding: " << prev_start << " " << feature[i].end + 1 
-    //        << " " << prev_counter << endl;
-  } 
+      (make_pair(feature[i].name, prev_counter )) );
+    // cout << "Adding: " << prev_start << " " << feature[i].end + 1
+    //      << " " << prev_counter << endl;
+  }
 
 }
 
-void 
+void
 Metagene::add_features(const string& bed_file) {
   BedReader reader(bed_file);
   GenomicRegion g;
   vector<string> fields;
- 
+
   vector<FeatureRegions> feature;
   reader.read_bed_line(g, fields);
   add_region(g, fields, feature);
 
   while(reader.read_bed_line(g, fields)) {
-    
+
     // check if the regions belongs to a new feature.
     // if it does, process the previous feature
     if (feature.back().name != fields[0]) {
       process_feature(feature);
-      ++n_features;   
+      ++n_features;
       // get rid of the old feature
       feature.clear();
     }
-    
+
     // if the region is part of an old feature, just keep track
     add_region(g, fields, feature);
   }
@@ -152,7 +152,7 @@ Metagene::add_features(const string& bed_file) {
 
 
 Metagene::Metagene(const string& bed_file, const size_t divisions) {
-  
+
   n_features = 0;
   n_divisions = divisions;
   add_features(bed_file);
@@ -162,38 +162,38 @@ Metagene::~Metagene() {
 
 }
 
-size_t 
+size_t
 Metagene::get_n_features() const {
   return n_features;
 }
-  
 
-void 
+
+void
 Metagene::get_feature_names(vector<string> &names) const {
   names.clear();
   names = feature_names;
 }
 
 
-void 
-Metagene::at(const GenomicRegion &in, 
-             vector<string> &feature, 
+void
+Metagene::at(const GenomicRegion &in,
+             vector<string> &feature,
              vector<size_t> &first, vector<size_t> &last) const {
 
   feature.clear();
   first.clear();
   last.clear();
 
-  unordered_map<string, size_t> regions;  
+  unordered_map<string, size_t> regions;
   size_t regions_counter = 0;
 
   // cout  << "querying " << in << endl;
 
-  vector<pair<GenomicRegion, FeatureVector<pair<string, size_t>>>> 
+  vector<pair<GenomicRegion, FeatureVector<pair<string, size_t>>>>
     feature_out;
   metagene.at(in, feature_out, true);
 
-  for(size_t i = 0; i < feature_out.size(); ++i) {
+  for (size_t i = 0; i < feature_out.size(); ++i) {
     // cout << feature_out[i].first << "\t";
     for (size_t j = 0; j < feature_out[i].second.size(); ++j) {
       // cout << feature_out[i].second.at(j).first << "\t";
@@ -210,10 +210,10 @@ Metagene::at(const GenomicRegion &in,
         first.push_back(feature_out[i].second.at(j).second);
         last.push_back(feature_out[i].second.at(j).second);
       }
-      // if the region already exists just upate the min and max 
+      // if the region already exists just upate the min and max
       else {
         // update the min and the max
-        // (only one will be updated depending on the direction). 
+        // (only one will be updated depending on the direction).
         if (feature_out[i].second.at(j).second < first[it->second]) {
           first[it->second] = feature_out[i].second.at(j).second;
         }
